+++ NEWS.md
@@ -3,6 +3,8 @@
 * xml and yaml objects are now stored in an R6 class called `yarn`.
 * testthat edition 3 is now being used with snapshot testing.
 * Tables are now pretty after a full loop `to_xml()` + `to_md()` (@pdaengeli, #9)
+* 2021-05-04: yarn objects remember the `sourcepos` and `encoding` options 
+  when resetting.
 
 # tinkr 0.0.0.9000